+++ src/dnsimple-test/ClientTest.cs
@@ -1,5 +1,7 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
 using dnsimple;
-using dnsimple.Services;
 using Moq;
 using NUnit.Framework;
 
@@ -24,34 +26,20 @@ namespace dnsimple_test
             Assert.That(_client.BaseUrl, Is.EqualTo("https://api.sandbox.dnsimple.com"));
         }
 
-        [Test]
-        public void HasHttpService()
-        {
-            Assert.That(_client.Http, Is.InstanceOf<HttpService>());
-        }
+        private static IEnumerable<PropertyInfo> ServiceProperties() =>
+            typeof(IClient).GetProperties()
+                .Where(property => property.PropertyType.Name.EndsWith("Service"));
 
-        [Test]
-        public void HasIdentityService()
+        [TestCaseSource(nameof(ServiceProperties))]
+        public void InitializesEveryService(PropertyInfo property)
         {
-            Assert.That(_client.Identity, Is.InstanceOf<IdentityService>());
-        }
-
-        [Test]
-        public void HasOAuthService()
-        {
-            Assert.That(_client.OAuth, Is.InstanceOf<OAuth2Service>());
+            Assert.That(property.GetValue(_client), Is.Not.Null);
         }
 
         [Test]
         public void HasVersion()
         {
             Assert.That(_client.Version, Is.EqualTo("v2"));
-        }
-
-        [Test]
-        public void ReturnsInstanceOfTheIdentityService()
-        {
-            Assert.That(_client.Identity, Is.InstanceOf<IdentityService>());
         }
 
         [Test]